テストケースの整理と保持する値をlongに変更。

+++ src/personal/mickie/education/calc/cpu/Formula.java
@@ -57,7 +57,7 @@ public class Formula {
 		lastTerm = new ValueKeySequence();
 	}
 
-	public int getResult() {
+	public long getResult() {
 		if (lastTerm.hasValue())
 			return lastTerm.getValue();
